requirements.txt
Flask
pandas
matplotlib
matplotlib_fontja
edinet_wrap
requests
python-dotenv
plotly